+++ tests/integration/cleanup-command.test.ts
@@ -50,7 +50,7 @@ jest.mock('../../src/phases/cleanup/artifact-cleaner.js', () => ({
   })),
 }));
 
-import * as fs from 'node:fs';
+import fs from 'fs-extra';
 import { findWorkflowMetadata } from '../../src/core/repository-utils.js';
 import { GitManager } from '../../src/core/git-manager.js';
 import { ArtifactCleaner } from '../../src/phases/cleanup/artifact-cleaner.js';